A7a — le chemin fusé cesse d'acheter une précision que sa référence n'a pas
La queue `KeepExact` était téléversée sur la carte en f32. Pas par choix de
précision : c'est la largeur de stockage du fichier, et `fused.rs` la
transportait faute de raison de faire autrement.

La question qui décidait de l'item est celle de la qualité, et elle a une
réponse dans le code, pas une intuition. Le bras dense auquel `fusedrun`
compare ses tokens charge par `sealed::load`, qui narrow la matrice
**entière** au dtype demandé — la queue est une plage de colonnes d'un
tenseur, rien ne l'exempte du `to_dtype`. Et les deux bras tournent en f16.
Donc la référence stocke déjà ces colonnes en binaire16 : la conversion
n'ajoute aucune perte, elle aligne le chemin fusé sur ce à quoi il se compare.
Argument plus fort trouvé en chemin : **aucun chiffre de qualité publié ne lit
ce tableau** — `ppl` et `mmlu` scorent tous deux par le chemin dense.

Comptabilité recomptée depuis les vraies formes, et la formule reproduit
d'abord les deux cellules publiées avant de prédire quoi que ce soit :
**5,162 → 5,094 b/param** au 4B, 33,91 Mo repris sur 16 957 440 poids de
queue. ⚠️ Le gain ne transporte pas — le 8B a `intermediate_size = 24 × 512`,
donc `down_proj` n'a aucune queue et il n'y reste que −0,039. C'est une
propriété des formes, pas de la méthode.

⚠️ Un décalage de comptabilité apparaît et il est étiqueté : les b/poids
noyau du banc facturent toujours la queue en f32, donc 4,729 n'est pas une
régression sur 4,804 — c'est une autre résidence. L'écart vaut exactement
`16 · tail_weights / weights` et la ligne de log de `fusedrun` l'imprime.

Treize mutants, treize tués. Il a fallu pour ça **réparer `h2f` côté hôte**,
qui rendait 0 sous licence « rien ne l'exécute » : `tail_dot_h` est de
l'arithmétique scalaire, donc le harnais clang++ l'exécute vraiment, et un
élargissement bouché aurait fait passer la sonde sur n'importe quelle queue.
La branche que NVRTC compile est inchangée, et élargir binary16 vers binary32
est exact — il n'y a rien à approximer.

Prouvé : la conversion, la comptabilité, et `tail_dot_h` lui-même, exécuté
contre une référence f64 et un rejeu f32. Non prouvé : les trois noyaux qui
l'appellent, dont la grille et les barrières n'existent que sur carte.

+++ llvq-cuda/src/gpu.rs
@@ -495,6 +495,14 @@ impl Cuda {
     /// input is: an inference runtime hands over candle's own
     /// `CudaSlice<half::f16>`, and this crate has no `half` dependency to name
     /// that type with. The kernel writes `unsigned short`, which is those bits.
+    ///
+    /// `tail` is `u16` for the same reason and the same encoding: since lot
+    /// A7a the f16-storing entry points hold the `KeepExact` tail as binary16,
+    /// which is what the dense arm they are diffed against holds it at. The
+    /// *bench* arms — `launch_slot`, `launch_slot_seg`, and planesbench's —
+    /// keep `f32`, because their published `b/poids noyau` bills 32 bits
+    /// there. Two residencies, two accountings, and the type keeps a caller
+    /// from handing one to the other.
     #[allow(clippy::too_many_arguments)]
     pub fn launch_slot_h<T: cudarc::driver::DeviceRepr>(
         &self,
@@ -504,7 +512,7 @@ impl Cuda {
         tab: &CudaSlice<u32>,
         gscale: &CudaSlice<f32>,
         rscale: &CudaSlice<f32>,
-        tail: &CudaSlice<f32>,
+        tail: &CudaSlice<u16>,
         x: &CudaSlice<f32>,
         y: &mut CudaSlice<T>,
         nblocks: u32,
